fix: merge with (#768);
