+++ .env
@@ -6,8 +6,5 @@ VITE_APP_API_BASE_URL=http://localhost:3002/
 # Whether long replies are supported, which may result in higher API fees
 VITE_GLOB_OPEN_LONG_REPLY=false
 
-# Default System Message
-VITE_DEFAULT_SYSTEM_MESSAGE=''
-
 # When you want to use PWA
 VITE_GLOB_APP_PWA=false
