+++ service/src/types.ts
@@ -9,7 +9,6 @@ export interface RequestProps {
 export interface ChatContext {
   conversationId?: string
   parentMessageId?: string
-  systemMessage?: string
 }
 
 export interface ChatGPTUnofficialProxyAPIOptions {
