+++ src/store/modules/index.ts
@@ -4,4 +4,3 @@ export * from './user'
 export * from './prompt'
 export * from './settings'
 export * from './auth'
-export * from './sysmsg'
