+++ src/store/modules/settings/helper.ts
@@ -2,14 +2,21 @@ import { ss } from '@/utils/storage'
 
 const LOCAL_NAME = 'settingsStorage'
 
+interface systemMessagePerChat {
+  uuid: number
+  message: string
+}
+
 export interface SettingsState {
   systemMessage: string
+  systemMessageAllChat: systemMessagePerChat[]
 }
 
 export function defaultSetting(): SettingsState {
   const currentDate = new Date().toISOString().split('T')[0]
   return {
     systemMessage: `You are ChatGPT, a large language model trained by OpenAI. Answer as concisely as possible.\nKnowledge cutoff: 2021-09-01\nCurrent date: ${currentDate}`,
+    systemMessageAllChat: [],
   }
 }
