+++ src/typings/env.d.ts
@@ -3,7 +3,6 @@
 interface ImportMetaEnv {
 	readonly VITE_GLOB_API_URL: string;
 	readonly VITE_APP_API_BASE_URL: string;
-    readonly VITE_DEFAULT_SYSTEM_MESSAGE: string;
 	readonly VITE_GLOB_OPEN_LONG_REPLY: string;
 	readonly VITE_GLOB_APP_PWA: string;
 }
